@@ -8,7 +8,10 @@
     "build:dev": "gulp dev",
     "server": "gulp dev-server",
     "start": "gulp dev-server",
-    "test": "mocha --require babel-core/register"
+    "stylelint": "stylelint src/**/*.scss src/**/*.css",
+    "eslint": "eslint src/**/*.js src/**/*.jsx",
+    "test": "mocha --compilers js:babel-core/register test/**/*.js",
+    "test:spec": "mocha --compilers js:babel-core/register test/helper.spec.js src/**/*.spec.js"
   },
   "license": "MIT",
   "repository": "https://github.com/elasim/website1-backend.git",
@@ -31,7 +34,7 @@
     "immutable": "^3.8.1",
     "jsonwebtoken": "^7.0.1",
     "lodash": "^4.14.1",
-    "material-ui": "^0.15.2",
+    "material-ui": "^0.15.3",
     "mysql": "^2.11.1",
     "ndjson": "^1.4.3",
     "nedb": "^1.8.0",
@@ -43,9 +46,9 @@
     "path-to-regexp": "^1.5.3",
     "pug": "^2.0.0-beta4",
     "raw-body": "^2.1.7",
-    "react": "^15.2.1",
-    "react-addons-css-transition-group": "^15.2.1",
-    "react-dom": "^15.2.1",
+    "react": "^15.3.0",
+    "react-addons-css-transition-group": "^15.3.0",
+    "react-dom": "^15.3.0",
     "react-intl": "^2.1.3",
     "react-redux": "^4.4.5",
     "react-router": "^2.5.2",
@@ -84,7 +87,9 @@
     "command-line-args": "^3.0.0",
     "command-line-usage": "^3.0.1",
     "css-loader": "^0.23.1",
+    "css-modules-require-hook": "^4.0.1",
     "del": "^2.2.1",
+    "enzyme": "^2.4.1",
     "eslint": "^2.12.0",
     "eslint-plugin-react": "^5.1.1",
     "event-source-polyfill": "0.0.7",
@@ -97,6 +102,7 @@
     "gulp-sourcemaps": "^1.6.0",
     "ignore-loader": "^0.1.1",
     "isomorphic-fetch": "^2.2.1",
+    "jsdom": "^9.4.1",
     "json-loader": "^0.5.4",
     "lodash": "^4.13.1",
     "mkdirp": "^0.5.1",
@@ -107,10 +113,15 @@
     "postcss-loader": "^0.9.1",
     "postcss-scss": "^0.1.8",
     "precss": "^1.4.0",
+    "react": "^15.3.0",
+    "react-addons-test-utils": "^15.3.0",
+    "react-dom": "^15.3.0",
     "react-hot-loader": "^1.3.0",
     "react-tap-event-plugin": "^1.0.0",
     "redux-logger": "^2.6.1",
     "sass-loader": "^3.2.1",
+    "sinon": "^1.17.5",
+    "sinon-chai": "^2.8.0",
     "style-loader": "^0.13.1",
     "stylelint": "^6.6.0",
     "stylelint-config-standard": "^9.0.0",
