@@ -1,5 +1,5 @@
-var webpack = require('webpack');
-var path = require('path');
+import webpack from 'webpack';
+import path from 'path';
 
 module.exports = {
 	context: path.join(__dirname, '../../src'),
